refactor: reorganized code
