+++ src/utils/wallet.ts
@@ -2,6 +2,8 @@ import { utils } from 'ethers'
 import { Seed } from './type'
 import { Utils } from '@ethersphere/bee-js'
 import { HD_PATH } from './account'
+import CryptoJS from 'crypto-js'
+import { bytesToWordArray, decrypt, encryptBytes, hexToWordArray } from './encryption'
 
 /**
  * Prepared data extract from seed to print
@@ -26,16 +28,16 @@ export function mnemonicToSeed(mnemonic: string): Seed {
  * Extract uncompressed public key from seed
  */
 export function uncompressedPublicKeyFromSeed(seed: Seed): string {
-  const hdNode = hdNodeFromSeed(seed)
+  const hdNode = mainHDNodeFromSeed(seed)
   const signingKey = new utils.SigningKey(hdNode.privateKey)
 
   return signingKey.publicKey
 }
 
 /**
- * Creates HDNode from seed
+ * Creates main HDNode account from seed
  */
-export function hdNodeFromSeed(seed: Seed): utils.HDNode {
+export function mainHDNodeFromSeed(seed: Seed): utils.HDNode {
   return utils.HDNode.fromSeed(seed).derivePath(HD_PATH)
 }
 
@@ -50,7 +52,7 @@ export function getSeedString(seed: Seed): string {
  * Gets prepared data for printing extracted from seed
  */
 export function getPrintDataFromSeed(seed: Seed): PrintSeedData {
-  const hdNode = hdNodeFromSeed(seed)
+  const hdNode = mainHDNodeFromSeed(seed)
 
   return {
     seed: getSeedString(seed),
@@ -58,4 +60,18 @@ export function getPrintDataFromSeed(seed: Seed): PrintSeedData {
     privateKey: hdNode.privateKey,
     address: hdNode.address,
   }
+}
+
+/**
+ * Encrypt seed with password and convert it to a string
+ */
+export function encryptSeed(seed: Seed, password: string): string {
+  return Utils.bytesToHex(encryptBytes(password, bytesToWordArray(seed)))
+}
+
+/**
+ * Decrypt seed string with password
+ */
+export function decryptSeedString(seed: string, password: string): Seed {
+  return Utils.hexToBytes(CryptoJS.enc.Hex.stringify(decrypt(password, hexToWordArray(seed))))
 }
